Split test code into multiple files
